+++ src/sprites/RocketSprite.ts
@@ -1,6 +1,6 @@
 import { Sprite, Texture } from "pixi.js";
 
-import { ROCKET_TEXTURE_PATH } from "../constants/path";
+import { ROCKET_TEXTURE_PATH } from "../constants/files";
 import { Coordinates } from "../types/coordinates";
 
 export class RocketSprite {
